Add first usable framegraph version, minor changes
