+++ src/arcana-incubator/pr-util/framegraph/fwd.hh
@@ -1,0 +1,17 @@
+#pragma once
+
+namespace inc::frag
+{
+class GraphCache;
+class GraphBuilder;
+
+struct virtual_resource_handle;
+struct access_mode;
+
+// internal structs
+struct virtual_resource;
+struct virtual_resource_version;
+struct internal_pass;
+struct physical_resource;
+struct guid_state;
+}
